ci(release): windows/amd (native-HIP) target + early version-collision guard (#358)
Adds the windows/amd quadrant to the release matrix as a native-HIP (gfx1151) target and an early version-collision guard that rejects a dispatched release whose version already matches the tree (previously that detonated at the publish git-commit after the full ~17-min matrix build).

- validate-inputs rejects new_version == the ref's [workspace.package].version, via a new release-notes.sh `current` subcommand (shared parser with `bump`, SSOT). Equality-only on purpose (sort -V orders pre-releases opposite to semver).
- windows-x86_64-amd-hip matrix target, experimental + continue-on-error. Compile-only in CI (hosted runners have no AMD GPU); the Windows HIP toolchain builds all kernels and links via stub libs. All fixes isolated to vendor="hip" so NVIDIA/SCALE/metal stay byte-identical.

+++ kernels/strix-hip/common/dense_gemm_tc.cu
@@ -79,10 +79,10 @@ extern "C" __global__ void dense_gemm_tc(
         // WMMA: A[16,16] × B^T[16,64] → C[16,64]. This warp does its 16×16 N-tile.
         v16bf a;
         #pragma unroll
-        for (int i = 0; i < 16; i++) a[i] = (__bf16)smem_A[lane_id & 15][i];
+        for (int i = 0; i < 16; i++) a[i] = (__bf16)(float)smem_A[lane_id & 15][i];
         v16bf b;
         #pragma unroll
-        for (int k = 0; k < 16; k++) b[k] = (__bf16)smem_B[k][n_warp_base + (lane_id & 15)];
+        for (int k = 0; k < 16; k++) b[k] = (__bf16)(float)smem_B[k][n_warp_base + (lane_id & 15)];
         acc = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc);
 
         __syncthreads();
