+++ kernels/strix-hip/common/inferspark_prefill.cu
@@ -203,7 +203,7 @@ extern "C" __global__ void inferspark_prefill(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
@@ -212,7 +212,7 @@ extern "C" __global__ void inferspark_prefill(
                     v16bf b;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        b[k] = (__bf16)smem_K[key_row][k_off + k];
+                        b[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc_s[nt]);
                 }
             }
@@ -305,7 +305,7 @@ extern "C" __global__ void inferspark_prefill(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
@@ -314,7 +314,7 @@ extern "C" __global__ void inferspark_prefill(
                     v16bf b;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        b[k] = (__bf16)smem_V[k_off + k][d_col];
+                        b[k] = (__bf16)(float)smem_V[k_off + k][d_col];
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc_o[nt]);
                 }
             }
@@ -501,7 +501,7 @@ extern "C" __global__ void inferspark_prefill_64(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
@@ -509,7 +509,7 @@ extern "C" __global__ void inferspark_prefill_64(
                     v16bf b;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        b[k] = (__bf16)smem_K[key_row][k_off + k];
+                        b[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc_s[nt]);
                 }
             }
@@ -589,7 +589,7 @@ extern "C" __global__ void inferspark_prefill_64(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
@@ -597,7 +597,7 @@ extern "C" __global__ void inferspark_prefill_64(
                     v16bf b;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        b[k] = (__bf16)smem_V[k_off + k][d_col];
+                        b[k] = (__bf16)(float)smem_V[k_off + k][d_col];
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, b, acc_o[nt]);
                 }
             }
