+++ kernels/strix-hip/common/prefill_paged_compute.cuh
@@ -230,7 +230,7 @@ extern "C" __global__ void KERNEL_NAME(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
@@ -238,7 +238,7 @@ extern "C" __global__ void KERNEL_NAME(
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[key_row][k_off + k];
+                        bb[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_s[nt]);
                 }
             }
@@ -326,7 +326,7 @@ extern "C" __global__ void KERNEL_NAME(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
 
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
@@ -334,7 +334,7 @@ extern "C" __global__ void KERNEL_NAME(
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_V[k_off + k][d_col];
+                        bb[k] = (__bf16)(float)smem_V[k_off + k][d_col];
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_o[nt]);
                 }
 #else
@@ -514,14 +514,14 @@ extern "C" __global__ void PAGED_CONCAT(KERNEL_NAME, _64)(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_Q[qk_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_Q[qk_m + lane_lo][k_off + i];
                 #pragma unroll
                 for (int nt = 0; nt < QK_N_TILES; nt++) {
                     unsigned int key_row = nt * 16 + lane_lo;
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_K[key_row][k_off + k];
+                        bb[k] = (__bf16)(float)smem_K[key_row][k_off + k];
                     acc_s[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_s[nt]);
                 }
             }
@@ -597,14 +597,14 @@ extern "C" __global__ void PAGED_CONCAT(KERNEL_NAME, _64)(
                 v16bf a;
                 #pragma unroll
                 for (int i = 0; i < 16; i++)
-                    a[i] = (__bf16)smem_P[pv_warp_m + lane_lo][k_off + i];
+                    a[i] = (__bf16)(float)smem_P[pv_warp_m + lane_lo][k_off + i];
                 #pragma unroll
                 for (int nt = 0; nt < PV_N_TILES; nt++) {
                     unsigned int d_col = (pv_n_start + nt) * 16 + lane_lo;
                     v16bf bb;
                     #pragma unroll
                     for (int k = 0; k < 16; k++)
-                        bb[k] = (__bf16)smem_V[k_off + k][d_col];
+                        bb[k] = (__bf16)(float)smem_V[k_off + k][d_col];
                     acc_o[nt] = __builtin_amdgcn_wmma_f32_16x16x16_bf16_w32(a, bb, acc_o[nt]);
                 }
 #else
